Reorganize event validation code (part 2)
